Fix linker errors on Linux.
