+++ Sources/Plasma/PubUtilLib/plAgeLoader/CMakeLists.txt
@@ -18,5 +18,8 @@ set(plAgeLoader_HEADERS
 
 add_library(plAgeLoader STATIC ${plAgeLoader_SOURCES} ${plAgeLoader_HEADERS})
 
+target_link_libraries(plAgeLoader CoreLib)
+target_link_libraries(plAgeLoader plResMgr)
+
 source_group("Source Files" FILES ${plAgeLoader_SOURCES})
 source_group("Header Files" FILES ${plAgeLoader_HEADERS})
