+++ Sources/Plasma/PubUtilLib/plResMgr/plPagePatcher.cpp
@@ -35,15 +35,14 @@ You can contact Cyan Worlds, Inc. by email legal@cyan.com
 #include "plRegistryHelpers.h"
 #include "plRegistryNode.h"
 #include "hsStream.h"
-
-#include "plAgeLoader/plResPatcher.h" // :(
+#include "plStatusLog/plStatusLog.h"
 
 #define KeyToKeyImpPublic(x) static_cast<plKeyImpPublic*>(static_cast<plKeyImp*>(x));
 
 #ifdef HS_DEBUGGING
 #   define PatcherLogDebug PatcherLog
 #else
-#   define PatcherLogDebug NULL_STMT
+#   define PatcherLogDebug (void)
 #endif
 
 class plKeyImpPublic : public plKeyImp
@@ -556,4 +555,34 @@ void plPagePatcher::WriteAndClear(plRegistryPageNode* const page, const plFileNa
     page->SetPagePathDirect(path);
     plWriteAndClearIterator iter;
     page->Write(&iter);
+}
+
+
+char* g_PatcherError = nullptr;
+
+void PatcherLog(PatcherLogType type, const char* format, ...)
+{
+    uint32_t color = 0;
+    switch (type)
+    {
+    case kHeader:       color = plStatusLog::kWhite;    break;
+    case kInfo:         color = plStatusLog::kBlue;     break;
+    case kMajorStatus:  color = plStatusLog::kYellow;   break;
+    case kStatus:       color = plStatusLog::kGreen;    break;
+    case kError:        color = plStatusLog::kRed;      break;
+    }
+
+    va_list args;
+    va_start(args, format);
+
+    plStatusLog* log = plStatusLogMgr::GetInstance().FindLog("patcher.log", true);
+    if (type == kError)
+    {
+        g_PatcherError = new char[1024]; // Deleted by Finish(false)
+        vsprintf(g_PatcherError, format, args);
+        log->AddLine(g_PatcherError, color);
+    } else
+        log->AddLineV(color, format, args);
+
+    va_end(args);
 }
