+++ Sources/Plasma/PubUtilLib/plStatusLog/CMakeLists.txt
@@ -16,7 +16,11 @@ set(plStatusLog_HEADERS
 )
 
 add_library(plStatusLog STATIC ${plStatusLog_SOURCES} ${plStatusLog_HEADERS})
-target_link_libraries(plStatusLog CoreLib plFile plUnifiedTime)
+
+target_link_libraries(plStatusLog CoreLib)
+target_link_libraries(plStatusLog pnTimer)
+target_link_libraries(plStatusLog plFile)
+target_link_libraries(plStatusLog plUnifiedTime)
 
 source_group("Source Files" FILES ${plStatusLog_SOURCES})
 source_group("Header Files" FILES ${plStatusLog_HEADERS})
