+++ Sources/Tools/plPatchUtil/CMakeLists.txt
@@ -6,70 +6,15 @@ include_directories("../../Plasma/PubUtilLib/inc")
 include_directories("../../Plasma/PubUtilLib")
 
 set(plPatchUtil_SOURCES
+    plAllCreatables.cpp
     main.cpp
 )
 
 add_executable(plPatchUtil ${plPatchUtil_SOURCES})
 
 target_link_libraries(plPatchUtil CoreLib)
-target_link_libraries(plPatchUtil plAgeDescription)
-target_link_libraries(plPatchUtil plAgeLoader)
-target_link_libraries(plPatchUtil plAudible)
-target_link_libraries(plPatchUtil plAudio)
-target_link_libraries(plPatchUtil plAudioCore)
-target_link_libraries(plPatchUtil plAvatar)
-target_link_libraries(plPatchUtil plCompression)
-target_link_libraries(plPatchUtil plContainer)
-target_link_libraries(plPatchUtil plDrawable)
-target_link_libraries(plPatchUtil plFile)
-target_link_libraries(plPatchUtil plGImage)
-target_link_libraries(plPatchUtil plGLight)
-target_link_libraries(plPatchUtil plInputCore)
-target_link_libraries(plPatchUtil plInterp)
-target_link_libraries(plPatchUtil plIntersect)
-target_link_libraries(plPatchUtil plMath)
-target_link_libraries(plPatchUtil plMessage)
-target_link_libraries(plPatchUtil plModifier)
-target_link_libraries(plPatchUtil plNetClient)
-target_link_libraries(plPatchUtil plNetClientComm)
-target_link_libraries(plPatchUtil plNetClientRecorder)
-target_link_libraries(plPatchUtil plNetCommon)
-target_link_libraries(plPatchUtil plNetGameLib)
-target_link_libraries(plPatchUtil plNetMessage)
-target_link_libraries(plPatchUtil plNetTransport)
-target_link_libraries(plPatchUtil plParticleSystem)
-target_link_libraries(plPatchUtil plPhysical)
-target_link_libraries(plPatchUtil plPhysX)
-target_link_libraries(plPatchUtil plPipeline)
-target_link_libraries(plPatchUtil plProgressMgr)
-target_link_libraries(plPatchUtil plResMgr)
-target_link_libraries(plPatchUtil plScene)
-target_link_libraries(plPatchUtil plSDL)
-target_link_libraries(plPatchUtil plSockets)
-target_link_libraries(plPatchUtil plStatGather)
-target_link_libraries(plPatchUtil plStatusLog)
-target_link_libraries(plPatchUtil plStreamLogger)
-target_link_libraries(plPatchUtil plSurface)
-target_link_libraries(plPatchUtil plTransform)
-target_link_libraries(plPatchUtil plUnifiedTime)
-target_link_libraries(plPatchUtil plVault)
-target_link_libraries(plPatchUtil pnAsyncCore)
-target_link_libraries(plPatchUtil pnAsyncCoreExe)
-target_link_libraries(plPatchUtil pnDispatch)
 target_link_libraries(plPatchUtil pnEncryption)
-target_link_libraries(plPatchUtil pnFactory)
-target_link_libraries(plPatchUtil pnGameMgr)
-target_link_libraries(plPatchUtil pnKeyedObject)
-target_link_libraries(plPatchUtil pnMessage)
-target_link_libraries(plPatchUtil pnModifier)
-target_link_libraries(plPatchUtil pnNetBase)
-target_link_libraries(plPatchUtil pnNetCli)
-target_link_libraries(plPatchUtil pnNetCommon)
-target_link_libraries(plPatchUtil pnNetProtocol)
-target_link_libraries(plPatchUtil pnNucleusInc)
-target_link_libraries(plPatchUtil pnSceneObject)
-target_link_libraries(plPatchUtil pnTimer)
-target_link_libraries(plPatchUtil pnUtils)
+target_link_libraries(plPatchUtil plResMgr)
 target_link_libraries(plPatchUtil pnUUID)
 
 if(USE_VLD)
